Conflicts:
	brogger/brog.go

+++ brogger/brog.go
@@ -1,6 +1,7 @@
 package brogger
 
 import (
+	"compress/gzip"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -20,18 +21,22 @@ import (
 // brog posts and watches for changes in posts and templates.
 type Brog struct {
 	*logMux
-	isProd   bool
-	Config   *Config
-	Pid      int
-	sock     string
-	tmplMngr *templateManager
-	postMngr *postManager
+	isProd      bool
+	Config      *Config
+	Pid         int
+	sock        string
+	tmplMngr    *templateManager
+	postMngr    *postManager
+	pageMngr    *postManager
+	middlewares [](func(http.HandlerFunc) http.HandlerFunc)
 }
 
 type appContent struct {
 	Posts     []*post
+	Pages     []*post
 	Languages []string
 	CurPost   *post
+	Redir     string
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -86,6 +91,10 @@ func (b *Brog) Close() error {
 		errHandler(b.postMngr.Close())
 	}
 
+	if b.pageMngr != nil {
+		errHandler(b.pageMngr.Close())
+	}
+
 	if b.tmplMngr != nil {
 		errHandler(b.tmplMngr.Close())
 	}
@@ -97,12 +106,19 @@ func (b *Brog) Close() error {
 		return fmt.Errorf("caught errors while closing, %v", errs)
 	}
 
-	if err := os.Remove(b.Config.PidFilename); err != nil {
-		return fmt.Errorf("deleting pidfile '%s', %v", b.Config.PidFilename, err)
+	if b.isProd {
+		if err := os.Remove(b.Config.PidFilename); err != nil {
+			return fmt.Errorf("deleting pidfile '%s', %v", b.Config.PidFilename, err)
+		}
 	}
 
-	if err := os.Remove(b.sock); err != nil {
-		return fmt.Errorf("deleting socket file: %v", err)
+	_, err := strconv.ParseInt(b.sock, 10, 0)
+	if nil != err {
+		if _, err := os.Stat(b.sock); nil == err {
+			if err := os.Remove(b.sock); err != nil {
+				return fmt.Errorf("deleting socket file: %v", err)
+			}
+		}
 	}
 	return nil
 }
@@ -136,13 +152,19 @@ func (b *Brog) ListenAndServe() error {
 		return fmt.Errorf("starting watchers, %v", err)
 	}
 
-	http.HandleFunc("/heartbeat", b.heartBeat) // don't log heartbeat, too noisy
-	http.HandleFunc("/changelang", b.logHandlerFunc(b.langSelectFunc))
-	http.HandleFunc("/posts/", b.logHandlerFunc(b.postFunc))
-	http.HandleFunc("/", b.logHandlerFunc(b.indexFunc))
+	b.HandleFunc("/heartbeat", b.heartBeat) // don't add middleware to heartbeat
+	b.middlewares = append(b.middlewares, b.logHandlerFunc)
+
+	// langSelect shouldn't have language middleware on it
+	b.HandleFunc("/changelang", b.langSelectFunc)
+	b.middlewares = append(b.middlewares, b.langHandlerFunc)
+
+	b.HandleFunc("/posts/", b.postFunc)
+	b.HandleFunc("/pages/", b.pageFunc)
+	b.HandleFunc("/", b.indexFunc)
 
 	fileServer := http.FileServer(http.Dir(b.Config.AssetPath))
-	http.Handle("/assets/", http.StripPrefix("/assets/", b.logHandler(fileServer)))
+	http.Handle("/assets/", http.StripPrefix("/assets/", b.logHandler(b.gzipHandler(fileServer))))
 
 	b.Ok("Assimilation completed.")
 	if b.isProd {
@@ -177,7 +199,15 @@ func (b *Brog) startWatchers() error {
 	if err != nil {
 		return fmt.Errorf("starting post manager, %v", err)
 	}
+
+	pageMngr, err := startPostManager(b, b.Config.PagePath)
+	if err != nil {
+		return fmt.Errorf("starting page manager, %v", err)
+	}
+
 	b.postMngr = postMngr
+	b.pageMngr = pageMngr
+
 	return nil
 }
 
@@ -208,6 +238,34 @@ func (b *Brog) logHandler(h http.Handler) http.Handler {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
+// Middleware
+////////////////////////////////////////////////////////////////////////////////
+
+func (b *Brog) HandleFunc(path string, h http.HandlerFunc) {
+	for _, middleware := range b.middlewares {
+		h = middleware(h)
+	}
+	http.HandleFunc(path, h)
+}
+
+//gzip handler for the assets files
+func (b *Brog) gzipHandler(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if !strings.Contains(req.Header.Get("Accept-Encoding"), "gzip") {
+			h.ServeHTTP(w, req)
+			return
+		}
+		w.Header().Set("Content-Encoding", "gzip")
+		gz := gzip.NewWriter(w)
+		defer func() {
+			_ = gz.Close()
+		}()
+		gzrw := gzipResponseWriter{Writer: gz, ResponseWriter: w}
+		h.ServeHTTP(gzrw, req)
+	})
+}
+
+////////////////////////////////////////////////////////////////////////////////
 // HandlerFuncs
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -219,15 +277,13 @@ func (b *Brog) heartBeat(rw http.ResponseWriter, req *http.Request) {
 
 func (b *Brog) indexFunc(rw http.ResponseWriter, req *http.Request) {
 
-	if b.Config.Multilingual {
-		b.langIndexFunc(rw, req)
-		return
-	}
-
-	posts := b.postMngr.GetAllPosts()
+	lang, _ := b.extractLanguage(req)
+	pages := b.pageMngr.GetAllPostsWithLanguage(lang)
+	posts := b.postMngr.GetAllPostsWithLanguage(lang)
 
 	data := appContent{
 		Posts:     posts,
+		Pages:     pages,
 		Languages: b.Config.Languages,
 		CurPost:   nil,
 	}
@@ -243,7 +299,10 @@ func (b *Brog) indexFunc(rw http.ResponseWriter, req *http.Request) {
 
 func (b *Brog) postFunc(rw http.ResponseWriter, req *http.Request) {
 
-	postID := path.Base(req.RequestURI)
+	lang, _ := b.extractLanguage(req)
+	pages := b.pageMngr.GetAllPostsWithLanguage(lang)
+
+	postID := path.Base(strings.SplitN(req.RequestURI, "?", 2)[0])
 	post, ok := b.postMngr.GetPost(postID)
 	if !ok {
 		b.Warn("not found, %v", req)
@@ -253,6 +312,7 @@ func (b *Brog) postFunc(rw http.ResponseWriter, req *http.Request) {
 
 	data := appContent{
 		Posts:     nil,
+		Pages:     pages,
 		Languages: b.Config.Languages,
 		CurPost:   post,
 	}
@@ -266,32 +326,39 @@ func (b *Brog) postFunc(rw http.ResponseWriter, req *http.Request) {
 	})
 }
 
-// Multilingual support
+func (b *Brog) pageFunc(rw http.ResponseWriter, req *http.Request) {
 
-func (b *Brog) langIndexFunc(rw http.ResponseWriter, req *http.Request) {
-	lang, validLang := b.extractLanguage(req)
-	b.setLangCookie(req, rw)
-	if !validLang {
-		b.langSelectFunc(rw, req)
+	lang, _ := b.extractLanguage(req)
+	pages := b.pageMngr.GetAllPostsWithLanguage(lang)
+
+	pageID := path.Base(strings.SplitN(req.RequestURI, "?", 2)[0])
+	page, ok := b.pageMngr.GetPost(pageID)
+
+	if !ok {
+		b.Warn("not found, %v", req)
+		http.NotFound(rw, req)
 		return
 	}
 
-	posts := b.postMngr.GetAllPostsWithLanguage(lang)
-
 	data := appContent{
-		Posts:     posts,
+		Posts:     nil,
+		Pages:     pages,
 		Languages: b.Config.Languages,
-		CurPost:   nil,
+		CurPost:   page,
 	}
 
-	b.tmplMngr.DoWithIndex(func(t *template.Template) {
+	b.tmplMngr.DoWithPost(func(t *template.Template) {
 		if err := t.Execute(rw, data); err != nil {
-			b.Err("serving index request, %v", err)
+			b.Err("serving post request for ID=%s, %v", pageID, err)
 			http.Error(rw, err.Error(), http.StatusInternalServerError)
 			return
 		}
 	})
 }
+
+////////////////////////////////////////////////////////////////////////////////
+// Multilingual support
+////////////////////////////////////////////////////////////////////////////////
 
 func (b *Brog) validLangInQuery(lang string) bool {
 	for _, val := range b.Config.Languages {
@@ -322,16 +389,24 @@ func (b *Brog) setLangCookie(req *http.Request, rw http.ResponseWriter) {
 	lang := req.URL.RawQuery
 	_, err := req.Cookie("lang")
 	if lang != "" && (err != nil || strings.HasSuffix(req.Referer(), "/changelang")) {
-		rw.Header().Add("Set-Cookie", "lang="+lang)
+		rw.Header().Add("Set-Cookie", "lang="+lang+";Path=/")
 	}
 }
 
 func (b *Brog) langSelectFunc(rw http.ResponseWriter, req *http.Request) {
 	b.Debug("Language not set for multilingual blog")
+
+	redirpath := req.URL.Path
+	if redirpath == "/changelang" {
+		redirpath = "/"
+	}
+
 	data := appContent{
 		Posts:     nil,
+		Pages:     nil,
 		Languages: b.Config.Languages,
 		CurPost:   nil,
+		Redir:     redirpath,
 	}
 
 	b.tmplMngr.DoWithLangSelect(func(t *template.Template) {
@@ -341,6 +416,20 @@ func (b *Brog) langSelectFunc(rw http.ResponseWriter, req *http.Request) {
 			http.Error(rw, err.Error(), http.StatusInternalServerError)
 			return
 		}
+	})
+}
+
+func (b *Brog) langHandlerFunc(h http.HandlerFunc) http.HandlerFunc {
+	return http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
+		if b.Config.Multilingual {
+			_, validLang := b.extractLanguage(req)
+			b.setLangCookie(req, rw)
+			if !validLang {
+				b.langSelectFunc(rw, req)
+				return
+			}
+		}
+		h.ServeHTTP(rw, req)
 	})
 }
